@@ -46,7 +46,11 @@ import { EstadoUsuariosServiceImpl } from '../../Core/Service/Implements/EstadoU
     ReactiveFormsModule,
     MatOptionModule,
     MatSelectModule,
-    HttpClientModule,
+    
+// TODO: `HttpClientModule` should not be imported into a component directly.
+// Please refactor the code to add `provideHttpClient()` call to the provider list in the
+// application bootstrap logic and remove the `HttpClientModule` import from this component.
+HttpClientModule,
     MatDatepickerModule,
     MatIconModule,
     MatAutocompleteModule,
