@@ -40,7 +40,11 @@ import { OriginInterceptor } from '../../Core/OriginInterceptor';
         FormsModule,
         CommonModule,
         ReactiveFormsModule,
-        HttpClientModule,
+        
+// TODO: `HttpClientModule` should not be imported into a component directly.
+// Please refactor the code to add `provideHttpClient()` call to the provider list in the
+// application bootstrap logic and remove the `HttpClientModule` import from this component.
+HttpClientModule,
         RouterLink,
         FormularioComponent,
         
