@@ -28,7 +28,11 @@ import { MatTableDataSource, MatTableModule } from '@angular/material/table';
     FormsModule,
     CommonModule,
     ReactiveFormsModule,
-    HttpClientModule,
+    
+// TODO: `HttpClientModule` should not be imported into a component directly.
+// Please refactor the code to add `provideHttpClient()` call to the provider list in the
+// application bootstrap logic and remove the `HttpClientModule` import from this component.
+HttpClientModule,
     ModalEditarComponent,
     MatTableModule
     
